\name{dbjumper}
\alias{dbjumper}

\title{ Parse mapping pathway }
\description{
parse mapping pathway for gene, protein, metabolomic
}
\usage{
dbjumper(id_list, from, to, pathMatrix)
}

\arguments{
  \item{id_list}{ }
  \item{from}{ }
  \item{to}{ }
  \item{pathMatrix}{  }
}

\value{
return a matrix with mapped result
}

\author{ Xiaoyong Sun }

\examples{

}
\keyword{ methods }

