\name{geneMap}
\alias{geneMap}

\title{ Mapping between NCBI}
\description{
map between ncbi ids
}
\usage{
geneMap(my_list, from_type, to_type)
}

\arguments{
  \item{my_list}{ }
  \item{from_type}{}
  \item{to_type}{}
}


\value{
return mapped result 
}

\author{ Xiaoyong Sun }

\examples{

}


\keyword{  methods  }

