\name{parse1}
\alias{parse1}

\title{ parse the returned result from NCBI }
\description{
parse the returned ID result from NCBI
}
\usage{
parse1(uri)
}
\arguments{
  \item{uri}{ the url address for parsing data}

}



\value{
retunr parsed id list
}

\author{Xiaoyong Sun }

\examples{

}

\keyword{  methods  }

