\name{processMessage}
\alias{processMessage}

\title{ Return a message}
\description{
return message
}
\usage{
processMessage(message)
}

\arguments{
  \item{message}{a character vector for warning message}

}

\value{
return a message
}

\author{ Xiaoyong Sun }

\examples{

}

\keyword{  methods  }

