\name{proteinMap}
\alias{proteinMap}

\title{ map among protein ids}
\description{
map among protein ids using ACC id as key
}
\usage{
proteinMap(my_list, from_type, to_type)
}

\arguments{
  \item{my_list}{ id list \code{my_list} here~~ }
  \item{from_type}{biological source type \code{from_type} here~~ }
  \item{to_type}{ destination source type \code{to_type} here~~ }
}

\value{
return mapped matrix
}

\author{ Xiaoyong Sun }

\examples{


}

\keyword{ methods }

