\name{validate}
\alias{validate}

\title{ Delete extra copies from matrix }
\description{
return unique ids
}
\usage{
validate(id_list)
}
\arguments{
  \item{id_list}{ the list of biological ids}

}


\value{
Return matrix of unique id pairs
}

\author{ Xiaoyong Sun }

\examples{

}

\keyword{ methods }

