\name{geneMap}
\alias{geneMap}

\title{ Mapping between NCBI}
\description{
map between ncbi ids
}
\usage{
geneMap(my_list, from_type, to_type)
}

\arguments{
  \item{my_list}{the list of biological ids }
  \item{from_type}{biological source type}
  \item{to_type}{destination source type}
}


\value{
return mapped result 
}

\author{ Xiaoyong Sun }

\examples{

}


\keyword{  methods  }

