\name{updateTerm}
\alias{updateTerm}
\alias{updateTerm}

\title{ Update Term}
\description{
when loading package, automatically get parameters from author's webpage
}
\usage{
updateTerm(biourl)
}

\arguments{
  \item{biourl}{a url for updating new configuration}
}

\author{Xiaoyng Sun }

\examples{

}

\keyword{ IO }

